Split the image file methods in another interface
