+++ FluentTerminal.App/Services/FileSystemService.cs
@@ -59,84 +59,6 @@ namespace FluentTerminal.App.Services
                 storageFile.Path);
         }
 
-        public async Task RemoveTemporaryBackgroundThemeImage()
-        {
-            var folder = await ApplicationData.Current
-                                     .LocalCacheFolder
-                                     .TryGetItemAsync("BackgroundThemeTmp");
-
-            if (folder == null)
-            {
-                return;
-            }
-
-            var backgroundThemeTmpFolder =
-                await ApplicationData.Current
-                                     .LocalCacheFolder
-                                     .GetFolderAsync(
-                                            "BackgroundThemeTmp");
-
-            await backgroundThemeTmpFolder.DeleteAsync();
-        }
-
-        public async Task<ImageFile> ImportTemporaryImageFile(IEnumerable<string> fileTypes)
-        {
-            var picker = new FileOpenPicker
-            {
-                SuggestedStartLocation = PickerLocationId.PicturesLibrary
-            };
-
-            foreach (var fileType in fileTypes)
-            {
-                picker.FileTypeFilter.Add(fileType);
-            }
-
-            var file = await picker.PickSingleFileAsync();
-
-            if (file != null)
-            {
-                var backgroundThemeTmpFolder = 
-                    await ApplicationData.Current
-                                         .LocalCacheFolder
-                                         .CreateFolderAsync(
-                                                "BackgroundThemeTmp", 
-                                                CreationCollisionOption.OpenIfExists);
-
-                var item = await backgroundThemeTmpFolder.TryGetItemAsync(file.Name);
-                
-                if (item == null)
-                {
-                    var storageFile = await file.CopyAsync(backgroundThemeTmpFolder, file.Name);
-                    
-                    return new ImageFile(
-                        storageFile.DisplayName, 
-                        storageFile.FileType, 
-                        $@"{backgroundThemeTmpFolder.Path}\{storageFile.Name}");
-                }
-
-                return new ImageFile(
-                    file.DisplayName,
-                    file.FileType,
-                    $@"{backgroundThemeTmpFolder.Path}\{item.Name}");
-            }
-
-            return null;
-        }
-
-        public async Task RemoveImportedImage(string fileName)
-        {
-            var backgroundThemeFolder = await ApplicationData.Current.RoamingFolder
-                .CreateFolderAsync("BackgroundTheme", CreationCollisionOption.OpenIfExists);
-
-            var item = await backgroundThemeFolder.TryGetItemAsync(fileName);
-
-            if (item != null)
-            {
-                var file = await backgroundThemeFolder.GetFileAsync(fileName);
-                await file.DeleteAsync(StorageDeleteOption.PermanentDelete);
-            }
-        }
-
         public async Task SaveTextFile(string name, string fileTypeDescription, string fileType, string content)
         {
             var picker = new FileSavePicker();
